src/services/terrain/SatelliteTerrainAnalyzer.ts
import { WeatherService } from '../weather/WeatherService';
import { TerrainAnalysisService } from './TerrainAnalysisService';
import { GeoPoint } from '@/types/geo';
import {
  TerrainConditions,
  TerrainFeature,
  TerrainHazard,
  SurfaceType,
  TerrainAnalysisResult
} from '@/types/terrain/types';

interface SatelliteImageryParams {
  center: GeoPoint;
  radius: number;
  resolution: 'ultra' | 'high' | 'medium' | 'low';
  timeRange?: {
    start: Date;
    end: Date;
  };
}

export class SatelliteTerrainAnalyzer {
  constructor(
    private weatherService: WeatherService,
    private terrainService: TerrainAnalysisService
  ) {}

  async analyzeTerrain(location: GeoPoint, weather: WeatherConditions): Promise<TerrainAnalysisResult> {
    try {
      const baseAnalysis = await this.terrainService.getTerrainConditions(location);
      const surfaceAnalysis = await this.analyzeSurfaceFromSatellite(location);
      const spectralAnalysis = await this.performSpectralAnalysis(location);
      const vegetationAnalysis = await this.analyzeVegetation(location);
      const waterAnalysis = await this.analyzeWaterBodies(location);
      
      const confidence = this.calculateConfidence(surfaceAnalysis, spectralAnalysis);

      // Create terrain analysis result
      const result: TerrainAnalysisResult = {
        surface: this.determineSurfaceType(surfaceAnalysis),
        hazards: this.identifyHazards(waterAnalysis, surfaceAnalysis),
        elevation: baseAnalysis.elevation,
        slope: baseAnalysis.slope,
        roughness: baseAnalysis.roughness,
        type: this.determineTerrainType(spectralAnalysis, vegetationAnalysis),
        features: this.extractFeatures(spectralAnalysis, vegetationAnalysis),
        confidence
      };

      return {
        ...result,
        risks: this.assessRisks(result, waterAnalysis),
        recommendations: this.generateRecommendations(result)
      };
    } catch (error) {
      console.error('Error in satellite terrain analysis:', error);
      throw new Error('Failed to analyze terrain from satellite data');
    }
  }

  private async fetchSatelliteData(params: SatelliteImageryParams): Promise<any> {
    // This would connect to a satellite imagery API
    // Placeholder implementation
    return {
      quality: 0.9,
      timestamp: new Date(),
      resolution: params.resolution === 'ultra' ? 0.5 : 
                 params.resolution === 'high' ? 1 :
                 params.resolution === 'medium' ? 5 : 10
    };
  }

  private async analyzeSpectralData(data: any) {
    // Analyze different spectral bands
    return {
      cloudCover: 0.1,
      bands: {
        visible: 0.8,
        nearInfrared: 0.7,
        thermal: 0.6
      }
    };
  }

  private async analyzeVegetation(data: any) {
    return {
      ndvi: 0.6, // Normalized Difference Vegetation Index
      density: 0.7,
      health: 0.8,
      type: 'mixed_forest'
    };
  }

  private async analyzeWaterBodies(data: any) {
    return {
      bodies: [],
      moisture: 0.5,
      floodRisk: 0.1
    };
  }

  private async analyzeSurfaceComposition(data: any) {
    return {
      primary: 'paved' as SurfaceType,
      coverage: 0.9,
      properties: {
        roughness: 0.3,
        stability: 0.8
      }
    };
  }

  private determineSurfaceType(analysis: any): SurfaceType {
    return analysis.primary;
  }

  private extractFeatures(spectral: any, vegetation: any): TerrainFeature[] {
    const features: TerrainFeature[] = [];
    
    if (vegetation.ndvi > 0.5) {
      features.push({
        type: 'vegetation',
        position: { latitude: 0, longitude: 0 }, // Would be calculated from actual data
        dimensions: {
          width: 100,
          height: 20
        },
        properties: {
          density: vegetation.density
        }
      });
    }

    return features;
  }

  private identifyHazards(water: any, surface: any): TerrainHazard[] {
    const hazards: TerrainHazard[] = [];

    if (water.floodRisk > 0.7) {
      hazards.push('flooding_risk');
    }

    if (surface.properties.stability < 0.3) {
      hazards.push('loose_surface');
    }

    return hazards;
  }

  private assessSurfaceQuality(surface: any, weather: any) {
    return {
      grip: surface.properties.roughness,
      stability: surface.properties.stability,
      drainage: 1 - (surface.properties.moisture || 0),
      maintenance: 0.9, // Would be calculated from actual maintenance data
      wear: 0.1,
      predictedDegradation: this.calculateDegradation(surface, weather)
    };
  }

  private calculateDegradation(surface: any, weather: any): number {
    let degradation = 0.01; // Base rate

    if (weather.conditions.includes('rain')) {
      degradation += 0.02;
    }

    if (surface.properties.stability < 0.5) {
      degradation += 0.03;
    }

    return degradation;
  }

  private calculateConfidence(params: {
    satelliteQuality: number;
    cloudCover: number;
    baseConfidence: number;
    resolution: string;
  }): number {
    let confidence = params.baseConfidence;

    // Adjust based on satellite data quality
    confidence *= (1 - params.cloudCover);
    confidence *= params.satelliteQuality;

    // Adjust based on resolution
    if (params.resolution === 'ultra') confidence *= 1;
    else if (params.resolution === 'high') confidence *= 0.9;
    else if (params.resolution === 'medium') confidence *= 0.8;
    else confidence *= 0.7;

    return Math.max(0.1, Math.min(1, confidence));
  }

  private calculateValidityPeriod(conditions: TerrainConditions, weather: any): Date {
    const validity = new Date();
    
    // Shorter validity for dynamic conditions
    if (conditions.hazards.includes('flooding_risk')) {
      validity.setHours(validity.getHours() + 1);
    } else if (weather.conditions.includes('rain')) {
      validity.setHours(validity.getHours() + 3);
    } else {
      validity.setHours(validity.getHours() + 12);
    }
    
    return validity;
  }

  private assessRisks(conditions: TerrainConditions, water: any) {
    return conditions.hazards.map(hazard => ({
      type: hazard,
      probability: hazard === 'flooding_risk' ? water.floodRisk : 0.5,
      impact: this.calculateHazardImpact(hazard),
      mitigations: this.getHazardMitigations(hazard)
    }));
  }

  private calculateHazardImpact(hazard: TerrainHazard): number {
    switch (hazard) {
      case 'flooding_risk':
        return 0.8;
      case 'loose_surface':
        return 0.5;
      default:
        return 0.3;
    }
  }

  private getHazardMitigations(hazard: TerrainHazard): string[] {
    switch (hazard) {
      case 'flooding_risk':
        return ['avoid_area', 'wait_for_water_level_decrease', 'use_alternate_route'];
      case 'loose_surface':
        return ['reduce_speed', 'use_appropriate_equipment', 'follow_marked_paths'];
      default:
        return ['proceed_with_caution'];
    }
  }

  private generateRecommendations(conditions: TerrainAnalysisResult) {
    return {
      maintenance: {
        priority: this.calculateMaintenancePriority(conditions),
        type: this.determineMaintenanceType(conditions),
        deadline: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000) // Default to 7 days from now
      },
      routing: {
        avoid: conditions.hazards.length > 0,
        alternatives: []
      }
    };
  }

  private calculateMaintenancePriority(conditions: TerrainAnalysisResult): number {
    // Implement priority calculation logic based on conditions
    return 1;
  }

  private determineMaintenanceType(conditions: TerrainAnalysisResult): string {
    // Implement maintenance type determination logic based on conditions
    return 'routine_maintenance';
  }
} 